+++ adapters/Snap/SnapAdapter/GADMediationSnapRewarded.h
@@ -1,4 +1,4 @@
-// Copyright 2019 Google LLC
+// Copyright 2021 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,10 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import <Foundation/Foundation.h>
+#import <GoogleMobileAds/GoogleMobileAds.h>
 
-@interface VungleRouterConfiguration : NSObject
-+ (void)setPublishIDFV:(BOOL)publish;
-+ (void)setMinSpaceForInit:(int)size;
-+ (void)setMinSpaceForAdLoad:(int)size;
+@interface GADMediationSnapRewarded : NSObject <GADMediationRewardedAd>
+
+- (void)renderRewardedAdForAdConfiguration:
+            (nonnull GADMediationRewardedAdConfiguration *)adConfiguration
+                         completionHandler:
+                             (nonnull GADMediationRewardedLoadCompletionHandler)completionHandler;
+
 @end