+++ adapters/Snap/SnapAdapter/GADMediatonAdapterSnapUtils.h
@@ -1,4 +1,4 @@
-// Copyright 2019 Google LLC
+// Copyright 2021 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,15 +13,7 @@
 // limitations under the License.
 
 #import <Foundation/Foundation.h>
-#import <GoogleMobileAds/GoogleMobileAds.h>
 
-@interface GADMAdapterVungleRewardedAd : NSObject <GADMediationRewardedAd>
-
-- (nonnull instancetype)
-    initWithAdConfiguration:(nonnull GADMediationRewardedAdConfiguration *)adConfiguration
-          completionHandler:(nonnull GADMediationRewardedLoadCompletionHandler)handler;
-- (nonnull instancetype)init NS_UNAVAILABLE;
-
-- (void)requestRewardedAd;
-
-@end
+/// Safely adds |object| to |set| if |object| is not nil.
+void GADMediationAdapterSnapMutableSetAddObject(NSMutableSet *_Nullable set,
+                                                NSObject *_Nonnull object);