+++ adapters/Snap/SnapAdapter/GADMediatonAdapterSnapUtils.m
@@ -1,4 +1,4 @@
-// Copyright 2019 Google LLC
+// Copyright 2021 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import <Foundation/Foundation.h>
+#import "GADMediatonAdapterSnapUtils.h"
 
-@interface VungleRouterConfiguration : NSObject
-+ (void)setPublishIDFV:(BOOL)publish;
-+ (void)setMinSpaceForInit:(int)size;
-+ (void)setMinSpaceForAdLoad:(int)size;
-@end
+void GADMediationAdapterSnapMutableSetAddObject(NSMutableSet *_Nullable set,
+                                                  NSObject *_Nonnull object) {
+  if (object) {
+    [set addObject:object];  // Allow pattern.
+  }
+}