+++ adapters/Facebook/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Facebook iOS Mediation Adapter Changelog
 
+#### [Version 6.9.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/facebook/FacebookAdapter-6.9.0.0.zip)
+- Verified compatibility with FAN SDK 6.9.0.
+- For Objective-C apps only, you must now add Swift paths to your target's `Build Settings` to prevent compile errors. See the [developer documentation](https://developers.google.com/admob/ios/mediation/facebook#step_4_additional_code_required) for more details.
+
+Built and tested with
+- Google Mobile Ads SDK version 8.12.0.
+- FAN SDK version 6.9.0.
+
 #### [Version 6.8.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/facebook/FacebookAdapter-6.8.0.0.zip)
 - Verified compatibility with FAN SDK 6.8.0.
 - Now requires minimum iOS version 10.0.